Develop

+++ color_utils.f90
@@ -4,8 +4,8 @@ module color_utils
   
 contains
   function MakeARGB(A, R, G, B) result(color)
-    integer(int), intent(in) :: A, R, G, B
-    integer(int) :: color
+    integer(int32), intent(in) :: A, R, G, B
+    integer(int32) :: color
     color = IOR(ISHFT(A, 24), IOR(ISHFT(R, 16), IOR(ISHFT(G, 8), B)))
   end function MakeARGB
